+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 1.5.2
 
 - Update Juggernaut effect bonuses
-- Spell modifier root nodes are now in mutual exclusion
+- Spell modifier root nodes are now mutually exclusive
+- Reworked Flame Slash cooldown modifier
+- Reworked Rain of Arrows cooldown modifier
 
 # 1.5.1